Update Reaction.cc
We were taking the energy in the CD and adding the energy lost in the dead layer (correct) and then adding the energy lost in the target after the interaction. Usually, we expect the gamma rays to be emitted in flight after they have left the target and not at the v/c corresponding to the point immediately after the interaction. So we don't want that correction. Even if if it does decay within the target, it will still probably have traversed some of the target, so this energy correction is almost never right. So let's remove it.